Kept only ci_sanity tests for policy detailed tests
